# 8274648: Improve logic for acquiring by reference parameters in downcall handles